requirements.txt
wikipedia-api
wikipedia
sentence-transformers
faiss-cpu
openai
numpy
